Add a new Burgernav component.
As you and all of developers knows, one of the best thing in native design is navbar. One of the popular way is to create burger type navbar and write all of the navbar content, with that way user can navigate easier in any mobile devices, so we create new navbar component where user experience in mobile devices ease. Hope you enjoy with this update :)